Various improvements

+++ SmartImage/Searching/Engines/TraceMoe/TraceMoeResult.cs
@@ -12,13 +12,13 @@ namespace SmartImage.Searching.Engines.TraceMoe
 	{
 		public double from { get; set; }
 		public double to { get; set; }
-		public int anilist_id { get; set; }
+		public long anilist_id { get; set; }
 		public double at { get; set; }
 		public string season { get; set; }
 		public string anime { get; set; }
 		public string filename { get; set; }
 
-		public int? episode { get; set; }
+		public long? episode { get; set; }
 
 		public string tokenthumb { get; set; }
 		public double similarity { get; set; }
@@ -27,7 +27,7 @@ namespace SmartImage.Searching.Engines.TraceMoe
 		public string title_chinese { get; set; }
 		public string title_english { get; set; }
 		public string title_romaji { get; set; }
-		public int mal_id { get; set; }
+		public long mal_id { get; set; }
 		public List<string> synonyms { get; set; }
 		public List<object> synonyms_chinese { get; set; }
 		public bool is_adult { get; set; }
@@ -36,15 +36,15 @@ namespace SmartImage.Searching.Engines.TraceMoe
 	[UsedImplicitly(ImplicitUseTargetFlags.WithMembers)]
 	public class TraceMoeRootObject
 	{
-		public int RawDocsCount { get; set; }
+		public long RawDocsCount { get; set; }
 		public long RawDocsSearchTime { get; set; }
-		public int ReRankSearchTime { get; set; }
+		public long ReRankSearchTime { get; set; }
 		public bool CacheHit { get; set; }
-		public int trial { get; set; }
-		public int limit { get; set; }
-		public int limit_ttl { get; set; }
-		public int quota { get; set; }
-		public int quota_ttl { get; set; }
+		public long trial { get; set; }
+		public long limit { get; set; }
+		public long limit_ttl { get; set; }
+		public long quota { get; set; }
+		public long quota_ttl { get; set; }
 		public List<TraceMoeDoc> docs { get; set; }
 	}
 }
